Expand state to track more things

+++ src/modules/logs/content/log_event_content/engineer_craft_event.rs
@@ -27,7 +27,7 @@ pub struct EngineerCraftEvent {
 #[serde(rename_all = "PascalCase")]
 pub struct EngineerCraftEventIngredient {
     pub name: Material,
-    pub count: u8,
+    pub count: u16,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
