+++ src/modules/logs/content/log_event_content/mission_completed_event.rs
@@ -116,5 +116,5 @@ pub struct MissionCompletedEventCommodityReward {
 pub struct MissionCompletedEventMaterialsReward {
     pub name: MixedMaterial,
     pub category: MixedMaterialCategory,
-    pub count: u8,
+    pub count: u16,
 }
